refactor: Tab, Article 공통 컴포넌트 수정 (#63)
* refactor: article 아이콘 정렬 + line-clamp 오류 수정 + 유틸 함수 분리

* refactor: 제거된 Input 공통 컴포넌트 제거 + max-width를 가지도록 수정

* refactor: tab 컴포넌트 width -\> maxWidth로 수정

---------
