+++ src/stories/Tab.stories.tsx
@@ -1,8 +1,8 @@
 import { Meta } from '@storybook/react';
 import { BsFire } from 'react-icons/bs';
 import { MdStars } from 'react-icons/md';
+import Article from '@components/Article';
 import ErrorText from '@components/ErrorText';
-import Input from '@components/Input';
 import Loader from '@components/Loader';
 import Tab from '@components/Tab';
 import { TabConstants } from '@constants/Tab';
@@ -17,7 +17,7 @@ export default meta;
 
 export const ArticleTab = () => {
   return (
-    <Tab>
+    <Tab maxWidth="25.875">
       <Tab.Item title={`${TabConstants.NEWEST}`} index="item1" width="8.625">
         <Loader />
       </Tab.Item>
@@ -27,7 +27,14 @@ export const ArticleTab = () => {
         icon={<BsFire className="w-[1.5rem] h-[1.5rem]" />}
         width="8.625"
       >
-        <Input placeholder="값을 입력해보세요" />
+        <Article
+          title="되겠지?"
+          nickname="@wukdddang"
+          postedDate="2023-08-29T09:28:39.390Z"
+          hasImage={false}
+          likes={10}
+          comments={42}
+        />
       </Tab.Item>
       <Tab.Item
         title={`${TabConstants.SUBSCRIBED}`}
@@ -43,12 +50,19 @@ export const ArticleTab = () => {
 
 export const TitleAndNicknameTab = () => {
   return (
-    <Tab>
+    <Tab maxWidth="23.375">
       <Tab.Item title={`${TabConstants.ARTICLE_TITLE}`} index="item1" width="11.6875">
         <Loader />
       </Tab.Item>
       <Tab.Item title={`${TabConstants.NICKNAME}`} index="item2" width="11.6875">
-        <Input placeholder="값을 입력해보세요" />
+        <Article
+          title="되겠지?"
+          nickname="@wukdddang"
+          postedDate="2023-08-29T09:28:39.390Z"
+          hasImage={false}
+          likes={10}
+          comments={42}
+        />
       </Tab.Item>
     </Tab>
   );
@@ -56,12 +70,19 @@ export const TitleAndNicknameTab = () => {
 
 export const SubscribeTab = () => {
   return (
-    <Tab>
+    <Tab maxWidth="23.375">
       <Tab.Item title={`${TabConstants.SUBSCRIBING}`} index="item1" width="11.6875">
         <Loader />
       </Tab.Item>
       <Tab.Item title={`${TabConstants.SUBSCRIBER}`} index="item2" width="11.6875">
-        <Input placeholder="값을 입력해보세요" />
+        <Article
+          title="되겠지?"
+          nickname="@wukdddang"
+          postedDate="2023-08-29T09:28:39.390Z"
+          hasImage={false}
+          likes={10}
+          comments={42}
+        />
       </Tab.Item>
     </Tab>
   );
